requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
python-dotenv

pydantic[email]
passlib[bcrypt]

pytest
requests

pytest-cov
python-multipart

python-jose
jinja2

alembic
aiofiles

black